Update for 1.0-SNAPSHOT with RoboVM support
Several modifications to be as close as possible to what has been done
for the Gradle support with the additions of some improvements:

- The Android project tries to follow the new Gradle project structure
except for the 'AndroidManifest.xml' that stays at the root of the
project to avoid problems with some IDEs with Android support enabled
such as NetBeans.

- The iOS project now relies on RoboVM but may still need some
corrections.

- The GWT project now includes the SuperDev mode and has a better
handling of the Jetty Maven plugin to avoid unnecessary recompilations.